+++ src/geometries/PrimitiveSphere.h
@@ -11,21 +11,21 @@
 class PrimitiveSphere : public Geometry {
 private:
   mathernogl::Vector3D centre;
-  float radius;
-  float radiusSquared;
+  double radius;
+  double radiusSquared;
   ContinousSamplerHelper samplerHelper;
 
 public:
   PrimitiveSphere();
-  PrimitiveSphere(const mathernogl::Vector3D& centre, float radius);
+  PrimitiveSphere(const mathernogl::Vector3D& centre, double radius);
 
   void setCentre(const mathernogl::Vector3D& centre);
-  void setCentre(float x, float y, float z);
-  void setRadius(float radius);
+  void setCentre(double x, double y, double z);
+  void setRadius(double radius);
 
   const mathernogl::Vector3D& getCentre() const;
-  float getRadius() const;
-  float getRadiusSquared() const;
+  double getRadius() const;
+  double getRadiusSquared() const;
   Vector2D getTextureCoord(const Vector3D& position) const;
 
   virtual bool hitTest(const Ray* ray, double* hitTValue, SurfaceInfo* surfaceInfo) const override;
